:recycle: Refactor font handling in EPUB reader components
- Moved SUPPORTED_FONT_OPTIONS to a new utility file for better organization.
- Updated FontFamily, FontSelect, and DefaultFontFamily components to import font options from the new utility file.
- This aims to better align font names with their css style conterparts
