+++ packages/browser/src/components/readers/epub/controls/FontFamily.tsx
@@ -4,7 +4,7 @@ import { isSupportedFont } from '@stump/sdk'
 import { useCallback } from 'react'
 
 import { useBookPreferences } from '@/scenes/book/reader/useBookPreferences'
-import { SUPPORTED_FONT_OPTIONS } from '@/scenes/settings/app/appearance/FontSelect'
+import { SUPPORTED_FONT_OPTIONS } from '@/utils/fonts'
 
 import { useEpubReaderContext } from '../context'
