+++ packages/browser/src/scenes/settings/app/appearance/FontSelect.tsx
@@ -4,6 +4,7 @@ import { isSupportedFont } from '@stump/sdk'
 import { useCallback } from 'react'
 
 import { usePreferences } from '@/hooks/usePreferences'
+import { SUPPORTED_FONT_OPTIONS } from '@/utils/fonts'
 
 /**
  * A component that allows the user to select the font for the app from a list of
@@ -55,14 +56,3 @@ export default function FontSelect() {
 }
 
 const localeKey = 'settingsScene.app/appearance.sections.fontSelect'
-
-export const SUPPORTED_FONT_OPTIONS = [
-	{ label: 'Atkinson Hyperlegible', value: 'atkinsonhyperlegible' },
-	{ label: 'Bitter', value: 'bitter' },
-	{ label: 'Charis SIL', value: 'charis' },
-	{ label: 'Inter', value: 'inter' },
-	{ label: 'Libre Baskerville', value: 'librebaskerville' },
-	{ label: 'Literata', value: 'literata' },
-	{ label: 'Nunito', value: 'nunito' },
-	{ label: 'OpenDyslexic', value: 'opendyslexic' },
-]
